Implement adapter for different cities and achive first output

+++ api/src/main/java/ar/edu/itba/pod/tpe2/query1/Q1ReducerFactory.java
@@ -3,23 +3,23 @@ package ar.edu.itba.pod.tpe2.query1;
 import com.hazelcast.mapreduce.Reducer;
 import com.hazelcast.mapreduce.ReducerFactory;
 
-public class Q1ReducerFactory implements ReducerFactory<String, Long, Long> {
+public class Q1ReducerFactory implements ReducerFactory<String, Integer, Integer> {
 
     @Override
-    public Reducer<Long, Long> newReducer(String s) {
+    public Reducer<Integer, Integer> newReducer(String s) {
         return new Q1Reducer();
     }
 
-    private static class Q1Reducer extends Reducer<Long, Long> {
-        private long sum = 0;
+    private static class Q1Reducer extends Reducer<Integer, Integer> {
+        private Integer sum = 0;
 
         @Override
-        public void reduce(Long value) {
+        public void reduce(Integer value) {
             sum += value;
         }
 
         @Override
-        public Long finalizeReduce() {
+        public Integer finalizeReduce() {
             return sum;
         }
     }
